seperated both research and ultimate faq
