+++ Admin/src/App.jsx
@@ -3,8 +3,8 @@ import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import Body from './Components/Body';
 import CreateBlog from './pages/CreateBlog';
 import User from './Pages/UserList/UserList';
-import Faq from './Pages/Faq/Faq';
 import Home from './pages/Home/home';
+import FAQManager from './Pages/Faq/FaqManager';
 
 function App() {
   return (
@@ -14,7 +14,7 @@ function App() {
         {/* Main app structure */}
         <Route path="/" element={<Body/>}>
           <Route path="/" element={<Home/>} />
-          <Route path="/createfaq" element={<Faq/>} />
+          <Route path="/createfaq" element={<FAQManager/>} />
           <Route path="/createblog" element={<CreateBlog />} />
           <Route path="/users" element={<User/>}> </Route>
          </Route>
